Refactor: ビジネスロジックをモデルからサービスクラスに分離
`src/day_trade/models/stock.py`の`Trade.get_portfolio_summary`メソッドに含まれていたポートフォリオサマリー計算ロジックを、新しく作成した`src/day_trade/core/portfolio.py`内の`PortfolioManager`クラスに移動しました。

- `Trade`モデルは純粋なデータ表現に責務を限定し、ビジネスロジックを`PortfolioManager`に分離しました。
- これにより、コードの凝集度と保守性が向上し、各コンポーネントの責務がより明確になりました。

Fixes #350
